graphics funtional
