+++ Frontend/open-data-portal/src/pages/Database.js
@@ -3,7 +3,6 @@ import { Box, Button, AppBar, Toolbar, Typography } from '@mui/material';
 import { Link } from 'react-router-dom';
 import MainTable1 from '../components/MainTable1';
 import MainTable2 from '../components/MainTable2';
-import MainTable3 from '../components/MainTable3';
 import MainTable4 from '../components/MainTable4';
 import MainTable5 from '../components/MainTable5';
 import Dataset from "../dashboard/DashboardContainer";
